Marketplace Implementation
Marketplace now in game when you select an empty.

For some reason when you press on the chicken it thinks your clikcing on the pig button.

think its cause the image is too big maybe?

Need to work on a way to get fillers in, eg Carrot, pigs wheat etc.

There is a fill type enum that can be used to check which button was chosen but they do not yet place anything.