Lockable areas and boundaries
Added in lockable areas using a blur on an object that sits above the map.
They get destroyed when the farmhouse is upgraded per level.

Boundaries for the camera have been added, should be asy to convert it to touch when that gets finalised.

Couple new assets in the build aswell.

Events.txt is now in its own folder GameData for when it gets built for other devices. and the path to it has been updated.

Camera has been moved to start in the correct location whre the farmhouse is.

//Tried having new prefabs for the lockable area but then decided no to use it and instead go with the blurred look