Added Radial Stats Menu
Shows:
Money that is earned through this field currently.
Food Earned.
Pollution of next field.
Upgrade Cost.

Fixed some values in resources due to them not being in all three of the levels....

Added a bunch of gets and sets all throughout the code to help with getting values.

Also added Battery Farm in code to setup for them coming in soon.