Added stats menu values
Changed distribution menu to switch sides when off/on screen.

Updated animation to accomodate this.

Save values of fields to transfer over to end creen.

Added stats menus that now work for both distribution screen and end screen.

Shows:
How many of each field
   - The cumulative money and food gain from each.

 - Totals of:
    Food produced, Money earned, Money Spent, Time Played, People Fed

Added battery farm radial menu script

Added more saving values

BUG FIX: Total food was not geting updated as there was a check for if the tutorial was the current scene instead of not the current scene.

Added button at end screen to allow for the stats to be toggled on and off instead of just on.